Build on Windows, Linux and macOS in CI
One workflow per platform, each building the game from a clean checkout, so that a
change which only compiles on one of them cannot pass unnoticed. Neither Windows nor
Linux had ever compiled this tree; both do now.

The shader toolchain decides what each job has to do. Windows needs nothing installed:
the SDK's own dxc.exe produces the DXIL the D3D12 backend loads. Linux downloads
Microsoft's prebuilt DirectXShaderCompiler. macOS has no prebuilt DXC at all, so a cold
build compiles that LLVM fork from source and the job caches the result. Linux builds
with clang: GCC takes -fms-extensions but does not bring __super with it, and the engine
uses that keyword in some 180 files.

Getting there took the engine off the pre-standard C++ it was still resting on. Every
module's CMakeLists had a WIN32 branch feeding it a different set of files; there is one
source list now, and Windows compiles what everyone else compiles. What it used to build
instead cannot build 64-bit at all — DirectPlay is gone from the Windows SDK, Bink drew
through d3d9x, the crash handler read CONTEXT.Eip, kdw's mini-windows.h had UINT_PTR as
a 32-bit int — so those follow DirectSound, the Win32 GUI helpers and the CD-key check
out of the build, and stay in the tree as reference.

The rest was inline x86 assembly (which MSVC accepts on no 64-bit target, and whose
portable branches were guarded to the non-Windows side), the DLL export macros keyed off
_DLL (MSVC's flag for linking the DLL *runtime*, not for building one), and a long tail
of things a conforming compiler refuses that clang only warns about: the address of a
temporary, `register`, string literals bound to char*.

Along the way it turned up bugs that were never about portability: containers streaming
their size() as 8 bytes into a 4-byte read (the replay format among them — the wire
fields say int32_t now), the vendored zlib compiling against the system's zlib.h, and a
`#pragma comment(lib)` with no library name that crashes MSVC and that clang had been
emitting into every object file as a malformed linker option.

PORTING.md records the traps in full, and which clang warnings predict them.

+++ Network/UniverseX.cpp
@@ -1,4 +1,5 @@
-#include "StdAfx.h"
+#include "stdafx.h"
+#include <cstdint>	// int32_t, for the 4-byte wire fields below
 #include "UniverseX.h"
 #include "RenderObjects.h"
 #include "Runtime.h"
@@ -13,7 +14,7 @@
 #include "Serialization/StringTable.h"
 
 #include "Lmcons.h"
-#include "vmap.h"
+#include "VMAP.H"
 #include "SelectManager.h"
 #include "NetCommands.h"
 #include "UserInterface/UI_Logic.h"
@@ -221,7 +222,7 @@ bool UniverseX::loadPlayReel(const char* fname)
 	temp.loadReplayInfoInBuf(buffer);
 
 	string name;
-	int varsSize;
+	int32_t varsSize;	// 4-byte wire field, written by savePlayReel()
 	buffer > varsSize;
 	currentProfileIntVariables_.clear();
 	for(; varsSize; varsSize--){
@@ -314,14 +315,17 @@ bool UniverseX::savePlayReel(const char* _fname)
 	mission.saveReplay(buffer);
 
 
-	buffer < currentProfileIntVariables_.size();
+	// int32_t, not size_t: this is the replay's wire format, and it is 4 bytes — what
+	// the original 32-bit build wrote and what loadPlayReel() reads back. On LP64
+	// size_t is 8 bytes, so this was streaming 8 into a 4-byte read.
+	buffer < (int32_t)currentProfileIntVariables_.size();
 	IntVariables::iterator i;
 	FOR_EACH(currentProfileIntVariables_, i)
 		buffer < StringOutWrapper(i->first) < i->second;
 
 	currentProfileParameters_.write(buffer);
 
-	buffer < voiceFileDurations_.size();
+	buffer < (int32_t)voiceFileDurations_.size();	// 4-byte wire field, as above
 	VoiceFileDurations::iterator vi;
 	FOR_EACH(voiceFileDurations_, vi)
 		buffer < StringOutWrapper(vi->first) < vi->second;
